[udev][usb-xhci][usb-hub] More work on enumerating devices attached at boot
Removed root hub port resetting at xhci driver initialization.
It turned out this was unnecessary, as resetting the controller
via asserting HCRST already resets the ports.
The real problem was that I didn't understand the difference between how
USB 2.0 and 3.0 root hub ports are reset.
For USB 3.0 hubs the reset state changes do not always include a change
to the "connected" bit.
So now the xhci root hub and usb-hub driver will handle the USB 3.0
reset path correctly.

Also added a mutex to the xhci root hub code to ensure that the reading and writing
of the status and status changed bits is thread safe.

Change-Id: I59e302fdebfe3d04aafb31fe4b1bad68ed2b4136